Fix kernel WebSocket protocol
