+++ tests/services/kernels/test_connection.py
@@ -5,12 +5,12 @@ from unittest.mock import MagicMock
 from jupyter_client.jsonutil import json_clean, json_default
 from jupyter_client.session import Session
 from tornado.httpserver import HTTPRequest
-from tornado.websocket import WebSocketHandler
 
 from jupyter_server.serverapp import ServerApp
 from jupyter_server.services.kernels.connection.channels import (
     ZMQChannelsWebsocketConnection,
 )
+from jupyter_server.services.kernels.websocket import KernelWebsocketHandler
 
 
 async def test_websocket_connection(jp_serverapp):
@@ -19,10 +19,11 @@ async def test_websocket_connection(jp_serverapp):
     kernel = app.kernel_manager.get_kernel(kernel_id)
     request = HTTPRequest("foo", "GET")
     request.connection = MagicMock()
-    handler = WebSocketHandler(app.web_app, request)
+    handler = KernelWebsocketHandler(app.web_app, request)
     handler.ws_connection = MagicMock()
     handler.ws_connection.is_closing = lambda: False
     conn = ZMQChannelsWebsocketConnection(parent=kernel, websocket_handler=handler)
+    handler.connection = conn
     await conn.prepare()
     conn.connect()
     await asyncio.wrap_future(conn.nudge())
@@ -37,7 +38,7 @@ async def test_websocket_connection(jp_serverapp):
     conn.handle_incoming_message(data)
     conn.handle_outgoing_message("iopub", session.serialize(msg))
     assert (
-        conn.select_subprotocol(["v1.kernel.websocket.jupyter.org"])
+        conn.websocket_handler.select_subprotocol(["v1.kernel.websocket.jupyter.org"])
         == "v1.kernel.websocket.jupyter.org"
     )
     conn.write_stderr("test", {})
